+++ lib/model_set.dart
@@ -1,0 +1,5 @@
+library model_set;
+
+export './src/model/DropdownModel.dart';
+export './src/model/LocalIdNameModel.dart';
+export './src/model/LocalReceivedNotification.dart';