+++ lib/src/Variant.dart
@@ -1,1 +1,0 @@
-enum Variant { dev, demo, prod }