+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+## 0.0.2
+* Remove error helper
+* Image helper
+  * with custom text color
+  * compress image with custom size
+* Add dropdown widget
+* Split lib to three set. helper_set, model_set, and widget_set
+
 ## 0.0.1
 * Initial